update product list

+++ web/ListAccount.jsp
@@ -19,6 +19,9 @@
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <title>List Account Page</title>
         <link rel="shortcut icon" href="images/web_logo.png">
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css" 
+              integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A==" 
+              crossorigin="anonymous" referrerpolicy="no-referrer" />
     </head>
 
     <body >
@@ -71,13 +74,13 @@
                             <c:choose>
                                 <c:when test="${login.account != a.account}">
                                     <a href="${updateUrl}" class="btn btn-warning">Update</a>
-                                    <a href="${isUseUpdateUrl}" class="btn btn-success" style="width: 90px">
+                                    <a href="${isUseUpdateUrl}" onclick="return confirm('Do you want to change status?')" class="btn btn-success" style="width: 90px">
                                         <c:choose>
                                             <c:when test="${a.isUse}">Deactive</c:when>
                                             <c:otherwise>Active</c:otherwise>
                                         </c:choose>
                                     </a>
-                                    <a href="${deleteUrl}" class="btn btn-danger" >Delete</a>
+                                    <a href="${deleteUrl}" onclick="return confirm('Do you want to delete this account?')" class="btn btn-danger" >Delete</a>
                                 </c:when>
                                 <c:otherwise>
                                     <a href="${updateUrl}" class="btn btn-warning disabled">Update</a>
@@ -87,7 +90,7 @@
                                             <c:otherwise>Active</c:otherwise>
                                         </c:choose>
                                     </a>
-                                    <a href="${deleteUrl}" class="btn btn-danger disabled" >Delete</a>
+                                    <a href="${deleteUrl}"  class="btn btn-danger disabled" >Delete</a>
                                 </c:otherwise>
                             </c:choose>
